Create PointChartDataModel abstraction; Create ScatterChartDataModel extension
